Separate the composition experiment from the original Nav proposal
